Initial commit
There are two projects here.  The first is a simple OSGi wrapper for
Jamod.  The second is the camel component.

For now, the camel component only implements a simple polling consumer
for the discrete inputs. Eventually, I hope to add support for
configuration (for example, specifying which inputs to poll), reading
coils and registers, and a producer which can be used to write values
to the modbus endpoint, as well.

The camel component builds, and the unit test for it works, but only if
you run a modbus simulator in a separate process.  For some reason, it
would get a "connection refused" when I ran the TCPListener from within
the unit test itself.