Making this not an OSGi bundle, and moving the packages around
It seems the preferred way of using Camel with OSGi is that the Camel
components themselves are not OSGi bundles, but instead you are
supposed to write an OSGi bundle with the camel route defined.

So, I've removed "impl" from the package namespaces and added a
META-INF/services/org/apache/camel/component/jamod file.  I've also
removed all the extra OSGi packaging stuff.